Bring back the triggers description and icons in form
